pyproject.toml
[project]
name = "clicker2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=4.1.1",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.2",
]
